requirements.txt
ezkl-gpu
onnx
torch
torchvision
lightning
hydra-core
psutil
pynvml
grpcio
grpcio-tools